Make all pointers in curl.h smart

+++ src/exceptions/request_except.cpp
@@ -2,6 +2,6 @@
 
 #include <string>
 
-FFS::BadHTTPStatusCode::BadHTTPStatusCode(int status):
+FFS::BadHTTPStatusCode::BadHTTPStatusCode(long status):
 	FFS::RequestException("Bad HTTP request, status code: " + std::to_string(status))
 {}
